adil.y-hr_management-2-hard-1756480596 (#5276)
You are Christopher Roach, a Hiring Manager, and your email is christopher.roach@yahoo.com. You want to publish the Design Director job in the Engineering department so that it moves from draft to open. You want to use the approval code POS0006 for the Hiring Manager approval and the approval code POS0003 for the HR Director approval. You also want the change to be documented so that the audit records show the previous status of draft and the new status of open. Additionally, check the audit is successfully done or not.